Refactor package structure
The new structure now allows to add package formats more easily.

This is necessary to support other package formats like .deb based packages.
