+++ notus/scanner/loader/loader.py
@@ -15,7 +15,7 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-from ..models.package import PackageAdvisories
+from ..models.packages.package import PackageAdvisories
 
 
 class AdvisoriesLoader:
