+++ notus/scanner/models/vulnerability.py
@@ -18,7 +18,7 @@
 from dataclasses import dataclass
 
 from ..models.advisory import Advisory
-from ..models.package import RPMPackage
+from .packages.package import Package
 
 
 @dataclass
@@ -27,6 +27,6 @@ class PackageVulnerability:
 
     host_ip: str
     host_name: str
-    package: RPMPackage
-    fixed_package: RPMPackage
+    package: Package
+    fixed_package: Package
     advisory: Advisory
