+++ tests/models/packages/test_package.py
@@ -17,15 +17,71 @@
 
 from dataclasses import FrozenInstanceError
 from unittest import TestCase
+from notus.scanner.errors import PackageError
 
-from notus.scanner.models.package import (
+from notus.scanner.models.packages.package import (
+    Package,
     AdvisoryReference,
     Architecture,
     OperatingSystemAdvisories,
     PackageAdvisories,
     PackageAdvisory,
-    RPMPackage,
 )
+from notus.scanner.models.packages.rpm import RPMPackage
+from notus.scanner.models.packages.deb import DEBPackage
+
+
+class PackageTestCase(TestCase):
+    def test_matching_hashes(self):
+        """hashes for the same package should match"""
+        package1 = Package(
+            name="foo-bar",
+            full_name="foo-bar-1.2.3-4.x86_64",
+            full_version="1.2.3-4.x86_64",
+        )
+        package2 = Package(
+            name="foo-bar",
+            full_name="foo-bar-1.2.3-4.x86_64",
+            full_version="1.2.3-4.x86_64",
+        )
+
+        self.assertEqual(hash(package1), hash(package2))
+
+    def test_not_matching_hashes(self):
+        """hashes for different packages should not match"""
+        package1 = Package(
+            name="foo-bar",
+            full_name="foo-bar-1.2.3-4.x86_64",
+            full_version="1.2.3-4.x86_64",
+        )
+        package2 = Package(
+            name="foo-bar",
+            full_name="foo-bar-1.2.4-4.x86_64",
+            full_version="1.2.4-4.x86_64",
+        )
+
+        self.assertNotEqual(hash(package1), hash(package2))
+
+    def test_gt_package_error(self):
+        package1 = RPMPackage(
+            name="foo-bar",
+            version="1.2.3",
+            release="4",
+            arch=Architecture.X86_64,
+            full_name="foo-bar-1.2.3-4.x86_64",
+            full_version="1.2.3-4.x86_64",
+        )
+        package2 = DEBPackage(
+            name="foo-bar",
+            epoch="1",
+            upstream_version="1.2.4",
+            debian_revision="4",
+            full_name="foo-bar-1.2.4-4.x86_64",
+            full_version="1.2.4-4.x86_64",
+        )
+        with self.assertRaises(PackageError):
+            if package1 > package2:
+                self.fail("PackageError should occur")
 
 
 class ArchitectureTestCase(TestCase):
@@ -44,261 +100,6 @@ class ArchitectureTestCase(TestCase):
         self.assertEqual(Architecture.I386, Architecture("i386"))
         self.assertEqual(Architecture.I686, Architecture("i686"))
         self.assertEqual(Architecture.X86_64, Architecture("x86_64"))
-
-
-class RPMPackageTestCase(TestCase):
-    def test_matching_hashes(self):
-        """hashes for the same package should match"""
-        package1 = RPMPackage(
-            name="foo-bar",
-            version="1.2.3",
-            release="4",
-            arch=Architecture.X86_64,
-            full_name="foo-bar-1.2.3-4.x86_64",
-            full_version="1.2.3-4.x86_64",
-        )
-        package2 = RPMPackage(
-            name="foo-bar",
-            version="1.2.3",
-            release="4",
-            arch=Architecture.X86_64,
-            full_name="foo-bar-1.2.3-4.x86_64",
-            full_version="1.2.3-4.x86_64",
-        )
-
-        self.assertEqual(hash(package1), hash(package2))
-
-    def test_not_matching_hashes(self):
-        """hashes for different packages should not match"""
-        package1 = RPMPackage(
-            name="foo-bar",
-            version="1.2.3",
-            release="4",
-            arch=Architecture.X86_64,
-            full_name="foo-bar-1.2.3-4.x86_64",
-            full_version="1.2.3-4.x86_64",
-        )
-        package2 = RPMPackage(
-            name="foo-bar",
-            version="1.2.4",
-            release="4",
-            arch=Architecture.X86_64,
-            full_name="foo-bar-1.2.4-4.x86_64",
-            full_version="1.2.4-4.x86_64",
-        )
-
-        self.assertNotEqual(hash(package1), hash(package2))
-
-    def test_compare_gt(self):
-        """packages should be comparable"""
-        package1 = RPMPackage(
-            name="foo-bar",
-            version="1.2.3",
-            release="4",
-            arch=Architecture.X86_64,
-            full_name="foo-bar-1.2.3-4.x86_64",
-            full_version="1.2.3-4.x86_64",
-        )
-        package2 = RPMPackage(
-            name="foo-bar",
-            version="1.2.4",
-            release="4",
-            arch=Architecture.X86_64,
-            full_name="foo-bar-1.2.4-4.x86_64",
-            full_version="1.2.4-4.x86_64",
-        )
-        self.assertGreater(package2, package1)
-
-        package2 = RPMPackage(
-            name="foo-bar",
-            version="1.2.3",
-            release="5",
-            arch=Architecture.X86_64,
-            full_name="foo-bar-1.2.3-5.x86_64",
-            full_version="1.2.3-5.x86_64",
-        )
-        self.assertGreater(package2, package1)
-
-    def test_compare_gt_different_architecture(self):
-        """packages of different architecture should not be comparable"""
-        package1 = RPMPackage(
-            name="foo-bar",
-            version="1.2.3",
-            release="4",
-            arch=Architecture.X86_64,
-            full_name="foo-bar-1.2.3-4.x86_64",
-            full_version="1.2.3-4.x86_64",
-        )
-        package2 = RPMPackage(
-            name="foo-bar",
-            version="1.2.3",
-            release="4",
-            arch=Architecture.AARCH64,
-            full_name="foo-bar-1.2.3-4.aarch64",
-            full_version="1.2.3-4.aarch64",
-        )
-        self.assertFalse(package2 > package1)
-        self.assertFalse(package1 > package2)
-
-    def test_compare_less(self):
-        """packages should be comparable"""
-        package1 = RPMPackage(
-            name="foo-bar",
-            version="1.2.3",
-            release="4",
-            arch=Architecture.X86_64,
-            full_name="foo-bar-1.2.3-4.x86_64",
-            full_version="1.2.3-4.x86_64",
-        )
-        package2 = RPMPackage(
-            name="foo-bar",
-            version="1.2.4",
-            release="4",
-            arch=Architecture.X86_64,
-            full_name="foo-bar-1.2.4-4.x86_64",
-            full_version="1.2.4-4.x86_64",
-        )
-        self.assertLess(package1, package2)
-
-        package2 = RPMPackage(
-            name="foo-bar",
-            version="1.2.3",
-            release="5",
-            arch=Architecture.X86_64,
-            full_name="foo-bar-1.2.3-5.x86_64",
-            full_version="1.2.3-5.x86_64",
-        )
-        self.assertLess(package1, package2)
-
-    def test_compare_equal(self):
-        """packages with the same data should be equal"""
-        package1 = RPMPackage(
-            name="foo-bar",
-            version="1.2.3",
-            release="4",
-            arch=Architecture.X86_64,
-            full_name="foo-bar-1.2.3-4.x86_64",
-            full_version="1.2.3-4.x86_64",
-        )
-        package2 = RPMPackage(
-            name="foo-bar",
-            version="1.2.3",
-            release="4",
-            arch=Architecture.X86_64,
-            full_name="foo-bar-1.2.3-4.x86_64",
-            full_version="1.2.3-4.x86_64",
-        )
-
-        self.assertEqual(package1, package2)
-
-    def test_from_full_name(self):
-        """it should be possible to create packages via the full name"""
-        self.assertIsNone(RPMPackage.from_full_name(None))
-
-        package = RPMPackage.from_full_name("keyutils-1.5.8-3.foo")
-        self.assertEqual(package.arch, Architecture.UNKNOWN)
-
-        package = RPMPackage.from_full_name("keyutils-1.5.8-3")
-        self.assertEqual(package.arch, Architecture.NOTSET)
-
-        package = RPMPackage.from_full_name(
-            "mesa-libgbm-11.2.2-2.20160614.x86_64"
-        )
-        self.assertEqual(package.arch, Architecture.X86_64)
-        self.assertEqual(package.name, "mesa-libgbm")
-        self.assertEqual(package.version, "11.2.2")
-        self.assertEqual(package.release, "2.20160614")
-        self.assertEqual(
-            package.full_name, "mesa-libgbm-11.2.2-2.20160614.x86_64"
-        )
-
-        package = RPMPackage.from_full_name("keyutils-1.5.8-3.x86_64")
-        self.assertEqual(package.arch, Architecture.X86_64)
-        self.assertEqual(package.name, "keyutils")
-        self.assertEqual(package.version, "1.5.8")
-        self.assertEqual(package.release, "3")
-        self.assertEqual(package.full_name, "keyutils-1.5.8-3.x86_64")
-
-        package = RPMPackage.from_full_name(
-            "httpd-manual-2.4.6-45.0.1.4.h10.noarch"
-        )
-        self.assertEqual(package.arch, Architecture.NOARCH)
-        self.assertEqual(package.name, "httpd-manual")
-        self.assertEqual(package.version, "2.4.6")
-        self.assertEqual(package.release, "45.0.1.4.h10")
-        self.assertEqual(
-            package.full_name, "httpd-manual-2.4.6-45.0.1.4.h10.noarch"
-        )
-
-        package = RPMPackage.from_full_name("cups-libs-1.6.3-26.h1.x86_64")
-        self.assertEqual(package.arch, Architecture.X86_64)
-        self.assertEqual(package.name, "cups-libs")
-        self.assertEqual(package.version, "1.6.3")
-        self.assertEqual(package.release, "26.h1")
-        self.assertEqual(package.full_name, "cups-libs-1.6.3-26.h1.x86_64")
-
-        package = RPMPackage.from_full_name("GConf2-3.2.6-8.x86_64")
-        self.assertEqual(package.arch, Architecture.X86_64)
-        self.assertEqual(package.name, "GConf2")
-        self.assertEqual(package.version, "3.2.6")
-        self.assertEqual(package.release, "8")
-        self.assertEqual(package.full_name, "GConf2-3.2.6-8.x86_64")
-
-        package = RPMPackage.from_full_name("libtool-ltdl-2.4.2-21.x86_64")
-        self.assertEqual(package.arch, Architecture.X86_64)
-        self.assertEqual(package.name, "libtool-ltdl")
-        self.assertEqual(package.version, "2.4.2")
-        self.assertEqual(package.release, "21")
-        self.assertEqual(package.full_name, "libtool-ltdl-2.4.2-21.x86_64")
-
-        package = RPMPackage.from_full_name("microcode_ctl-2.1-22.6.h2.x86_64")
-        self.assertEqual(package.arch, Architecture.X86_64)
-        self.assertEqual(package.name, "microcode_ctl")
-        self.assertEqual(package.version, "2.1")
-        self.assertEqual(package.release, "22.6.h2")
-        self.assertEqual(package.full_name, "microcode_ctl-2.1-22.6.h2.x86_64")
-
-        package = RPMPackage.from_full_name("postgresql-libs-9.2.23-3.x86_64")
-        self.assertEqual(package.arch, Architecture.X86_64)
-        self.assertEqual(package.name, "postgresql-libs")
-        self.assertEqual(package.version, "9.2.23")
-        self.assertEqual(package.release, "3")
-        self.assertEqual(package.full_name, "postgresql-libs-9.2.23-3.x86_64")
-
-        package = RPMPackage.from_full_name("NetworkManager-1.8.0-9.h2.x86_64")
-        self.assertEqual(package.arch, Architecture.X86_64)
-        self.assertEqual(package.name, "NetworkManager")
-        self.assertEqual(package.version, "1.8.0")
-        self.assertEqual(package.release, "9.h2")
-        self.assertEqual(package.full_name, "NetworkManager-1.8.0-9.h2.x86_64")
-
-        package = RPMPackage.from_full_name(
-            "perl-Pod-Escapes-1.04-285.h2.noarch"
-        )
-        self.assertEqual(package.arch, Architecture.NOARCH)
-        self.assertEqual(package.name, "perl-Pod-Escapes")
-        self.assertEqual(package.version, "1.04")
-        self.assertEqual(package.release, "285.h2")
-        self.assertEqual(
-            package.full_name, "perl-Pod-Escapes-1.04-285.h2.noarch"
-        )
-
-        package = RPMPackage.from_full_name(" libtool-ltdl-2.4.2-21.x86_64\r\n")
-        self.assertEqual(package.arch, Architecture.X86_64)
-
-    def test_from_name_and_full_version(self):
-        """it should be possible to create packages from name and full
-        version"""
-        self.assertIsNone(RPMPackage.from_name_and_full_version(None, None))
-
-        package = RPMPackage.from_name_and_full_version(
-            "cups-libs", "1.6.3-26.h1.x86_64"
-        )
-        self.assertEqual(package.arch, Architecture.X86_64)
-        self.assertEqual(package.name, "cups-libs")
-        self.assertEqual(package.version, "1.6.3")
-        self.assertEqual(package.release, "26.h1")
-        self.assertEqual(package.full_name, "cups-libs-1.6.3-26.h1.x86_64")
 
 
 class OperatingSystemAdvisoriesTestCase(TestCase):
