+++ tests/models/test_vulnerability.py
@@ -19,7 +19,8 @@ from datetime import datetime
 from unittest import TestCase
 
 from notus.scanner.models.advisory import Advisory, Severity
-from notus.scanner.models.package import Architecture, RPMPackage
+from notus.scanner.models.packages.package import Architecture
+from notus.scanner.models.packages.rpm import RPMPackage
 from notus.scanner.models.vulnerability import PackageVulnerability
 
 
@@ -69,14 +70,18 @@ class PackageVulnerabilityTestCase(TestCase):
 
         package = package_vulnerability.package
         self.assertEqual(package.name, "foo-bar")
-        self.assertEqual(package.version, "1.2.3")
-        self.assertEqual(package.release, "4")
+        self.assertEqual(package.version, "1.2.3")  # pylint: disable=no-member
+        self.assertEqual(package.release, "4")  # pylint: disable=no-member
         self.assertEqual(package.arch, Architecture.X86_64)
         self.assertEqual(package.full_name, "foo-bar-1.2.3-4.x86_64")
 
         fixed_package = package_vulnerability.fixed_package
         self.assertEqual(fixed_package.name, "foo-bar")
-        self.assertEqual(fixed_package.version, "1.2.4")
-        self.assertEqual(fixed_package.release, "4")
+        self.assertEqual(
+            fixed_package.version, "1.2.4"  # pylint: disable=no-member
+        )
+        self.assertEqual(
+            fixed_package.release, "4"  # pylint: disable=no-member
+        )
         self.assertEqual(fixed_package.arch, Architecture.X86_64)
         self.assertEqual(fixed_package.full_name, "foo-bar-1.2.4-4.x86_64")
